add back crack gpt
