+++ Cargo.toml
@@ -9,10 +9,10 @@ members = [
 resolver = "2"
 
 [workspace.dependencies]
-crack-core = { path = "crack-core", default-features = true, version = "0.2.12" }
-crack-osint = { path = "crack-osint", default-features = true, version = "0.1.1" }
-crack-config = { path = "crack-config", default-features = true, version = "0.1.0" }
-crack-gpt = { path = "crack-gpt", default-features = true, version = "0.1.0" }
+crack-core = { path = "crack-core", default-features = true, version = "0.2" }
+crack-osint = { path = "crack-osint", default-features = true, version = "0.1" }
+crack-config = { path = "crack-config", default-features = true, version = "0.1" }
+crack-gpt = { path = "crack-gpt", default-features = true, version = "0.1" }
 tungstenite = "^0.20.1"
 async-tungstenite = "^0.23.0"
 reqwest = { version = "0.11.22", features = ["blocking", "json"] }
