+++ crack-core/src/commands/crack_gpt/mod.rs
@@ -1,0 +1,3 @@
+pub mod crack_gpt;
+
+pub use crack_gpt::*;
