+++ crack-gpt/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
 name = "crack-gpt"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+version = "0.1.1"
+edition = "2021"
 
 [dependencies]
-chatgpt_rs = "1.2.3"
-crack-core = { path = "../crack-core", default-features = true }
+chatgpt_rs = "1.2"
+url = "2.2"
+tracing = "0.1"
+tokio = { version = "1.11", features = ["rt-multi-thread", "macros"] }
